requirements.txt
asgiref==3.2.3
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-debug-toolbar==2.2
django-embed-video==1.3.2
django-model-utils==4.0.0
django-vote==2.1.7
django-widget-tweaks==1.4.8
idna==2.8
oauthlib==3.1.0
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
urllib3==1.25.9
pylint-django==2.0.15
Pillow==7.1.2
ffmpeg-python==0.2.0
django-crispy-forms==1.9.1
django-admin-sortable2==0.7.6
xlsxwriter==1.2.8
django_extensions==2.2.9
dj-database-url==0.5.0
django-sendgrid-v5==0.8.1
psycopg2==2.8.5
wheel==0.34.2
django-import-export==2.2.0
safety==1.9.0
